requirements.txt
lightkurve>=2.4
astropy>=6.0
torch>=2.2
torchvision>=0.17
timm>=1.0
pennylane>=0.39
pyts>=0.13
scikit-learn>=1.4
wandb>=0.18
matplotlib>=3.8
numpy>=1.26
pandas>=2.1
scipy>=1.13
pyyaml>=6.0
pytest>=8.0
pytest-cov>=5.0
ruff>=0.6
pyvo>=1.4
